src/client-public/scripts/database/requirements.txt
python-dotenv=0.20.0
psycopg2=2.9.3
uuid=1.30
